Many of Edit Done

+++ database/migrations/2020_11_11_229889_create_categories_table.php
@@ -18,7 +18,7 @@ class CreateCategoriesTable extends Migration
             $table->bigIncrements('id');
             $table->bigInteger('admin_id')->unsigned()->nullable();
             $table->foreign('admin_id')->references('id')->on('admins')->onDelete('cascade');
-            $table->bigInteger('Parent_id')->nullable();
+            $table->unsignedInteger('Parent_id')->nullable();
             $table->string('name');
             $table->string('image_cate')->nullable();
 			$table->softDeletes();
